refactor(analytics): Share StatCard primitive across both analytics dashboards
The StatCard previously lived inline in AnalyticsDashboard. Hoist it
into its own file (src/components/recommendations/StatCard.tsx) and
have AdvancedDiscoveryAnalytics' SummaryCards consume it so both
top-level dashboards render with the same canonical .stat treatment:
font-display value, eyebrow header, font-mono trend, optional progress
bar / gradient / glow.

AdvancedDiscoveryAnalytics SummaryCards now show:
- Feedback (count, caption)
- Accept Rate (with week-over-week trend chip)
- Discovery (with progress bar)
- Monthly (with month-over-month trend chip)

Also bumps the 4-up grid breakpoint from lg to xl to match the
Overview tab fix from the prior commit.
